+++ R/target_sda.R
@@ -14,6 +14,7 @@
 #' @param ald An asset-level data frame like [r2dii.data::ald_demo].
 #' @param co2_intensity_scenario A scenario data frame like
 #'   [r2dii.data::co2_intensity_scenario_demo].
+#' @param region_isos A data frame like [r2dii.data::region_isos] (default).
 #' @param use_credit_limit Logical vector of length 1. `FALSE` defaults to using
 #'   the column `loan_size_outstanding`. Set to `TRUE` to instead use the column
 #'   `loan_size_credit_limit`.
@@ -21,11 +22,9 @@
 #'   `weighted_production_value` at the portfolio-level. Set to `TRUE` to output
 #'   `weighted_production_value` at the company-level.
 #'
-#' @return  A tibble with the CO2 emissions factors attributed to
-#' the portfolio. These values include the portfolio's actual projected CO2
-#' emissions factors, the scenario pathway CO2 emissions factors and the SDA
-#' calculated portfolio target emissions factors (see column
-#' `emission_factor_metric`).
+#' @return A tibble including the summarized columns `emission_factor_metric` and
+#'   `emission_factor_value`. If `by_company = TRUE`, the output will also have
+#'   the column `name_ald`.
 #'
 #' @export
 #'
@@ -77,7 +76,8 @@ target_sda <- function(data,
                        ald,
                        co2_intensity_scenario,
                        use_credit_limit = FALSE,
-                       by_company = FALSE) {
+                       by_company = FALSE,
+                       region_isos = r2dii.data::region_isos) {
   stopifnot(
     is.data.frame(data),
     is.data.frame(ald),
@@ -131,17 +131,41 @@ target_sda <- function(data,
   check_crucial_names(co2_intensity_scenario, crucial_scenario)
   walk_(crucial_scenario, ~ check_no_value_is_missing(co2_intensity_scenario, .x))
 
+  region_isos <- region_isos %>%
+    filter(source %in% unique(co2_intensity_scenario$scenario_source))
+
+  if (identical(nrow(region_isos), 0L)) {
+    warn("Found no matching regions for input scenario", class = "no_match")
+    return(empty_target_sda_output())
+  }
+
+  ald <- ald %>%
+    mutate(plant_location = tolower(.data$plant_location)) %>%
+    left_join(region_isos, by = c(plant_location = "isos")) %>%
+    rename(scenario_source = "source")
+
   ald_by_sector <- ald %>%
     aggregate_excluding(c("technology", "plant_location", "country_of_domicile"))
 
   data <- inner_join(data, ald_by_sector, by = ald_columns())
 
+  summary_groups <- c(
+    "region",
+    "scenario_source"
+  )
+
   if (by_company) {
     data <- data %>%
-      summarize_unweighted_emission_factor(!!!rlang::syms("name_ald"))
+      summarize_weighted_emission_factor(
+        !!!rlang::syms(c(summary_groups, "name_ald")),
+        use_credit_limit = use_credit_limit
+      )
   } else {
     data <- data %>%
-      summarize_weighted_emission_factor(use_credit_limit = use_credit_limit)
+      summarize_weighted_emission_factor(
+        !!!rlang::syms(summary_groups),
+        use_credit_limit = use_credit_limit
+      )
   }
 
   data <- data %>%
@@ -154,7 +178,7 @@ target_sda <- function(data,
 
   corporate_economy <- calculate_market_average(ald_by_sector)
 
-  interpolate_groups <- c("scenario", "sector", "region")
+  interpolate_groups <- c("scenario_source", "scenario", "sector", "region")
 
   interpolated_scenario <- interpolate_scenario_yearly(
     co2_intensity_scenario,
@@ -174,7 +198,7 @@ target_sda <- function(data,
   adjusted_scenario_with_p <- add_p_to_scenario(adjusted_scenario)
 
   target_summary_groups <- maybe_add_name_ald(
-    c("sector", "scenario"),
+    c("sector", "scenario", "region", "scenario_source"),
     by_company
   )
 
@@ -242,14 +266,14 @@ maybe_add_name_ald <- function(data, by_company = FALSE) {
 
 calculate_market_average <- function(data) {
   data %>%
-    group_by(.data$sector, .data$year) %>%
+    group_by(.data$sector, .data$year, .data$region, .data$scenario_source) %>%
     summarize(
       sector_total_production = sum(.data$production),
       # Alias emission_factor_corporate_economy
       .x = list(.data$production * .data$emission_factor)
     ) %>%
     unnest(cols = .data$.x) %>%
-    group_by(.data$sector, .data$year) %>%
+    group_by(.data$sector, .data$year, .data$region, .data$scenario_source) %>%
     summarize(.x = sum(.data$.x / .data$sector_total_production)) %>%
     rename(emission_factor_corporate_economy = .data$.x) %>%
     ungroup()
@@ -257,18 +281,28 @@ calculate_market_average <- function(data) {
 
 compute_ald_adjusted_scenario <- function(data, corporate_economy) {
   corporate_economy_baseline <- corporate_economy %>%
-    group_by(.data$sector) %>%
+    group_by(.data$sector, .data$scenario_source, .data$region) %>%
     filter(.data$year == min(.data$year, na.rm = TRUE)) %>%
     select(
       .data$sector,
+      .data$scenario_source,
+      .data$region,
       baseline_emission_factor = .data$emission_factor_corporate_economy
     ) %>%
     ungroup()
 
   data %>%
     filter(.data$year >= min(corporate_economy$year)) %>%
-    inner_join(corporate_economy_baseline, by = "sector") %>%
-    group_by(.data$scenario, .data$sector) %>%
+    inner_join(
+      corporate_economy_baseline,
+      by = c("sector", "scenario_source", "region")
+    ) %>%
+    group_by(
+      .data$scenario,
+      .data$sector,
+      .data$scenario_source,
+      .data$region
+    ) %>%
     arrange(.data$year) %>%
     mutate(
       baseline_adjustment =
@@ -280,6 +314,8 @@ compute_ald_adjusted_scenario <- function(data, corporate_economy) {
     select(
       .data$scenario,
       .data$sector,
+      .data$scenario_source,
+      .data$region,
       .data$year,
       .data$emission_factor_adjusted_scenario
     )
@@ -289,7 +325,12 @@ add_p_to_scenario <- function(data) {
   p <- function(x) (x - last(x)) / (first(x) - last(x))
 
   data %>%
-    group_by(.data$sector, .data$scenario) %>%
+    group_by(
+      .data$sector,
+      .data$scenario,
+      .data$region,
+      .data$scenario_source
+    ) %>%
     arrange(.data$year) %>%
     mutate(p = p(.data$emission_factor_adjusted_scenario)) %>%
     ungroup()
@@ -299,7 +340,10 @@ compute_loanbook_targets <- function(data,
                                      scenario_with_p,
                                      ...) {
   data %>%
-    right_join(scenario_with_p, by = c("year", "sector")) %>%
+    right_join(
+      scenario_with_p,
+      by = c("year", "sector", "region", "scenario_source")
+    ) %>%
     group_by(...) %>%
     arrange(.data$year) %>%
     mutate(