+++ NEWS.md
@@ -1,7 +1,7 @@
 # r2dii.analysis (development version)
 
-* `target_sda()` now outputs unweighted `emission_factor` by default, if 
-  `by_company` is `TRUE` (#394). 
+* `target_sda()` now outputs unweighted `emission_factor` if `by_company` is 
+  `TRUE` (#376). 
 
 * `target_market_share()` now outputs two new columns, `percentage_of_initial_production_by_scope` 
 and `scope` (ADO #4143).